requirements.txt
absl-py>=0.9.0
aiohttp>=3.6.2
alabaster>=0.7.12
alpha-vantage>=2.2.0
anaconda-client>=1.7.2
anaconda-navigator>=1.9.12
anaconda-project>=0.8.3
argh>=0.26.2
asgiref>=3.2.10
asn1crypto>=1.3.0
astroid>=2.3.3
astropy>=4.0
astunparse>=1.6.3
async-timeout>=3.0.1
atomicwrites>=1.3.0
attrs>=19.3.0
autobahn>=20.7.1
Automat>=20.2.0
autopep8>=1.4.4
Babel>=2.8.0
backcall>=0.1.0
backports.functools-lru-cache>=1.6.1
backports.shutil-get-terminal-size>=1.0.0
backports.tempfile>=1.0
backports.weakref>=1.0.post1
backtrader>=1.9.76.123
bcrypt>=3.1.7
beautifulsoup4>=4.8.2
bkcharts>=0.2
bleach>=3.1.0
bokeh>=1.4.0
boto>=2.49.0
Bottleneck>=1.3.2
bsedata>=0.4.0
cachetools>=4.1.1
certifi>=2019.11.28
cffi>=1.14.0
chardet>=3.0.4
Click>=7.0
cloudpickle>=1.3.0
clyent>=1.2.2
colorama>=0.4.3
comtypes>=1.1.7
conda>=4.8.2
conda-build>=3.18.11
conda-package-handling>=1.6.0
conda-verify>=3.4.2
constantly>=15.1.0
contextlib2>=0.6.0.post1
cryptography>=2.8
cycler>=0.10.0
Cython>=0.29.15
cytoolz>=0.10.1
dask>=2.11.0
dateutils>=0.6.8
decorator>=4.4.1
defusedxml>=0.6.0
diff-match-patch>=20181111
dill>=0.3.2
distributed>=2.11.0
Django>=3.0.8
docutils>=0.16
empyrical>=0.5.3
entrypoints>=0.3
enum34>=1.1.10
et-xmlfile>=1.0.1
fastcache>=1.1.0
fastnumbers>=3.0.0
filelock>=3.0.12
fix-yahoo-finance>=0.1.37
flake8>=3.7.9
Flask>=1.1.1
fsspec>=0.6.2
future>=0.18.2
gast>=0.3.3
gevent>=1.4.0
glob2>=0.7
google-auth>=1.18.0
google-auth-oauthlib>=0.4.1
google-pasta>=0.2.0
greenlet>=0.4.15
grpcio>=1.30.0
h5py>=2.10.0
HeapDict>=1.0.1
html5lib>=1.0.1
hyperlink>=19.0.0
hypothesis>=5.5.4
idna>=2.8
iexfinance>=0.4.3
imageio>=2.6.1
imagesize>=1.2.0
importlib-metadata>=1.5.0
incremental>=17.5.0
inflection>=0.5.0
install>=1.3.3
intervaltree>=3.0.2
ipykernel>=5.1.4
ipython>=7.12.0
ipython-genutils>=0.2.0
ipywidgets>=7.5.1
isort>=4.3.21
itsdangerous>=1.1.0
jdcal>=1.4.1
jedi>=0.14.1
Jinja2>=2.11.1
joblib>=0.14.1
json5>=0.9.1
jsonschema>=3.2.0
jupyter>=1.0.0
jupyter-client>=5.3.4
jupyter-console>=6.1.0
jupyter-core>=4.6.1
jupyterlab>=1.2.6
jupyterlab-server>=1.0.6
Keras>=2.4.3
Keras-Preprocessing>=1.1.2
keyring>=21.1.0
kiteconnect>=3.8.2
kiwisolver>=1.1.0
lazy-object-proxy>=1.4.3
libarchive-c>=2.8
llvmlite>=0.31.0
locket>=0.2.0
lxml>=4.5.0
Mako>=1.1.3
Markdown>=2.4.1
MarkupSafe>=1.1.1
matplotlib>=3.1.3
mccabe>=0.6.1
menuinst>=1.4.16
mistune>=0.8.4
mkl-fft>=1.0.15
mkl-random>=1.1.0
mkl-service>=2.3.0
mock>=4.0.1
more-itertools>=8.2.0
mpmath>=1.1.0
msgpack>=0.6.1
multidict>=4.7.6
multipledispatch>=0.6.0
multitasking>=0.0.9
navigator-updater>=0.2.1
nbconvert>=5.6.1
nbformat>=5.0.4
networkx>=2.4
nltk>=3.4.5
nose>=1.3.7
notebook>=6.0.3
nsepy>=0.8
nsetools>=1.0.11
numba>=0.48.0
numexpr>=2.7.1
numpy>=1.19.0
numpydoc>=0.9.2
oauthlib>=3.1.0
olefile>=0.46
openpyxl>=3.0.3
opt-einsum>=3.2.1
packaging>=20.1
pandas>=1.0.1
pandas-datareader>=0.9.0
pandocfilters>=1.4.2
paramiko>=2.7.1
parso>=0.5.2
partd>=1.1.0
path>=13.1.0
pathlib2>=2.3.5
pathtools>=0.1.2
patsy>=0.5.1
pdoc>=0.3.2
pep8>=1.7.1
pexpect>=4.8.0
pickleshare>=0.7.5
Pillow>=7.0.0
pkginfo>=1.5.0.1
pluggy>=0.13.1
ply>=3.11
prettytable>=0.7.2
prometheus-client>=0.7.1
prompt-toolkit>=3.0.3
protobuf>=3.12.2
psutil>=5.6.7
py>=1.8.1
py-lets-be-rational>=1.0.1
py-vollib>=1.0.1
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pycodestyle>=2.5.0
pycosat>=0.6.3
pycparser>=2.19
pycrypto>=2.6.1
pycurl>=7.43.0.5
pydocstyle>=4.0.1
pyflakes>=2.1.1
pyfolio>=0.9.2
Pygments>=2.5.2
PyHamcrest>=2.0.2
pylint>=2.4.4
PyNaCl>=1.3.0
pyodbc>==4.0.0-unsupported
pyOpenSSL>=19.1.0
pyparsing>=2.4.6
pyreadline>=2.1
pyrsistent>=0.15.7
PySocks>=1.7.1
pytest>=5.3.5
pytest-arraydiff>=0.3
pytest-astropy>=0.8.0
pytest-astropy-header>=0.1.2
pytest-doctestplus>=0.5.0
pytest-openfiles>=0.4.0
pytest-remotedata>=0.3.2
python-dateutil>=2.8.1
python-jsonrpc-server>=0.3.4
python-language-server>=0.31.7
pytrends>=4.7.3
pytz>=2019.3
PyWavelets>=1.1.1
pywin32>=227
pywin32-ctypes>=0.2.0
pywinpty>=0.5.7
PyYAML>=5.3
pyzmq>=18.1.1
QDarkStyle>=2.8
QtAwesome>=0.6.1
qtconsole>=4.6.0
QtPy>=1.9.0
Quandl>=3.5.0
quantrautil>=0.0.3
requests>=2.22.0
requests-oauthlib>=1.3.0
rope>=0.16.0
rsa>=4.6
Rtree>=0.9.3
ruamel-yaml>=0.15.87
scikit-image>=0.16.2
scikit-learn>=0.22.1
scipy>=1.4.1
seaborn>=0.10.0
Send2Trash>=1.5.0
service-identity>=18.1.0
simplegeneric>=0.8.1
simplejson>=3.17.0
singledispatch>=3.4.0.3
six>=1.14.0
snowballstemmer>=2.0.0
sortedcollections>=1.1.2
sortedcontainers>=2.1.0
soupsieve>=1.9.5
Sphinx>=2.4.0
sphinxcontrib-applehelp>=1.0.1
sphinxcontrib-devhelp>=1.0.1
sphinxcontrib-htmlhelp>=1.0.2
sphinxcontrib-jsmath>=1.0.1
sphinxcontrib-qthelp>=1.0.2
sphinxcontrib-serializinghtml>=1.1.3
sphinxcontrib-websupport>=1.2.0
spyder>=4.0.1
spyder-kernels>=1.8.1
SQLAlchemy>=1.3.13
sqlparse>=0.3.1
statsmodels>=0.11.0
stocker>=0.1.5
style>=1.1.0
sympy>=1.5.1
ta>=0.5.25
tables>=3.6.1
tabulate>=0.8.7
tblib>=1.6.0
tensorboard>=2.2.2
tensorboard-plugin-wit>=1.7.0
tensorflow>=2.2.0
tensorflow-estimator>=2.2.0
termcolor>=1.1.0
terminado>=0.8.3
testpath>=0.4.4
texttable>=1.6.2
toolz>=0.10.0
tornado>=6.0.3
tqdm>=4.42.1
traitlets>=4.3.3
Twisted>=20.3.0
txaio>=20.4.1
ujson>=1.35
unicodecsv>=0.14.1
update>=0.0.1
urllib3>=1.25.8
watchdog>=0.10.2
wcwidth>=0.1.8
webencodings>=0.5.1
Werkzeug>=1.0.0
widgetsnbextension>=3.5.1
win-inet-pton>=1.1.0
win-unicode-console>=0.5
wincertstore>=0.2
wrapt>=1.11.2
xlrd>=1.2.0
XlsxWriter>=1.2.7
xlwings>=0.17.1
xlwt>=1.3.0
xmltodict>=0.12.0
yahoo-fin>=0.8.5
yapf>=0.28.0
yarl>=1.4.2
yfinance>=0.1.54
zict>=1.0.0
zipp>=2.2.0
zope.interface>=5.1.0
